Separate Nexus registration tests
