+++ tests/Temporalio.Tests/Worker/NexusWorkerTests.cs
@@ -17,6 +17,7 @@ using Temporalio.Workflows;
 using Xunit;
 using Xunit.Abstractions;
 
+[CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
 public class NexusWorkerTests : WorkflowEnvironmentTestBase
 {
     public NexusWorkerTests(ITestOutputHelper output, WorkflowEnvironment env)
@@ -95,7 +96,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_SimpleService_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -121,7 +121,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ContextHasEndpoint()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -145,7 +144,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_SimpleWorkflow_Succeeds()
     {
         WorkflowRunOperationContext? capturedContext = null;
@@ -205,7 +203,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_WaitForeverWorkflow_CanBeCanceled()
     {
         // Build the worker options w/ the nexus service
@@ -274,35 +271,7 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
                 Client.GetWorkflowHandle(token.WorkflowId).GetResultAsync())).InnerException);
     }
 
-    [NexusService]
-    public interface IBadService
-    {
-        [NexusOperation]
-        int DoSomething(string name);
-    }
-
-    [NexusServiceHandler(typeof(IBadService))]
-    public class BadService
-    {
-        [NexusOperationHandler]
-        public IOperationHandler<string, string> DoSomething() =>
-            throw new NotImplementedException();
-    }
-
     [Fact]
-    public async Task ExecuteNexusOperationAsync_BadService_FailsRegistration()
-    {
-        var exc = Assert.Throws<ArgumentException>(() => new TemporalWorker(
-            Client,
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").AddNexusService(new BadService())));
-        Assert.Equal("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Equal(
-            "Expected return type of IOperationHandler<String, Int32>",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-    [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_SyncTimeout_FailsAsExpected()
     {
         var cancellationReasonSource = new TaskCompletionSource<string?>();
@@ -347,7 +316,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_RequestDeadline_SetOnContext()
     {
         var requestDeadlineSource = new TaskCompletionSource<DateTime?>();
@@ -371,7 +339,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationSummary_FoundInHistory()
     {
         string expectedSummary = "custom operation summary";
@@ -405,7 +372,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ScheduleToStartTimeout_FailsAsExpected()
     {
         var cancellationReasonSource = new TaskCompletionSource<string?>();
@@ -450,7 +416,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_StartToCloseTimeout_FailsAsExpected()
     {
         // Build a workflow-backed async operation that will never complete
@@ -498,7 +463,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_SimpleWorkflow_ConflictPolicy()
     {
         // Example of a Nexus service that creates a second operation with the same ID and
@@ -570,7 +534,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_HandlerSignal_ForwardsInboundLinks()
     {
         // A target workflow that the handler will signal from inside the operation. The signal it
@@ -623,7 +586,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_HandlerStart_ForwardsInboundLinks()
     {
         // A handler that plain-starts a target workflow from inside the operation. The start request
@@ -675,7 +637,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [SkippableFact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_HandlerSignal_PropagatesBacklink()
     {
         // The backward direction is gated by the server's history.enableCHASMSignalBacklinks dynamic
@@ -739,7 +700,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [SkippableFact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_HandlerSignalsMultiple_PropagatesAllBacklinks()
     {
         // Integration-level counterpart to the unit-level accumulation tests: a single operation
@@ -817,7 +777,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [SkippableFact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_AsyncHandlerSignal_PropagatesBacklinkOnStarted()
     {
         // Backlink propagation for an async operation: the handler signals a target workflow (the
@@ -891,7 +850,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_BadArgs_FailsOperation()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -913,7 +871,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_Untyped_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -929,7 +886,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_InputManip_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -950,7 +906,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ApplicationFailure_NonRetryable()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -976,7 +931,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationException_ProperlyFails()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -999,7 +953,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationExceptionWithCause_PreservesCauseChain()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1030,7 +983,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationExceptionCanceled_ProperlyFails()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1053,7 +1005,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationExceptionCanceledWithCause_PreservesCauseChain()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1082,7 +1033,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_HandlerException_ProperlyFails()
     {
         // Non-retryable
@@ -1107,7 +1057,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ManualDefinition_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1170,7 +1119,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_Interceptor_Reached()
     {
         var workflowId = $"wf-{Guid.NewGuid()}";
@@ -1205,7 +1153,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ServiceNotFound_ProperlyFails()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1228,7 +1175,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_OperationNotFound_ProperlyFails()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1316,7 +1262,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_VoidTypes_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -1424,7 +1369,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CancelWaitCompleted_ProperlyCancels()
     {
         var ret = await RunCancelOperationScenarioAsync(ICancelTypeService.Scenario.WaitCompletedTargetTimerAndRethrow);
@@ -1445,7 +1389,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CancelWaitRequested_ProperlyCancels()
     {
         var ret = await RunCancelOperationScenarioAsync(ICancelTypeService.Scenario.WaitRequestedTargetHang);
@@ -1465,7 +1408,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CancelWaitRequested_ProperlyFails()
     {
         // Confirm workflow fails trying to cancel
@@ -1478,7 +1420,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CancelAbandon_ProperlyCancels()
     {
         var ret = await RunCancelOperationScenarioAsync(ICancelTypeService.Scenario.Abandon);
@@ -1498,7 +1439,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CancelTryCancel_ProperlyCancels()
     {
         var ret = await RunCancelOperationScenarioAsync(ICancelTypeService.Scenario.TryCancelHandlerFail);
@@ -1596,7 +1536,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CodecFailure_IsRetried()
     {
         var codec = new FailOnceCodec();
@@ -1654,7 +1593,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_NoValueInputWithCodec_Succeeds()
     {
         var newOptions = (TemporalClientOptions)Client.Options.Clone();
@@ -1702,7 +1640,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ConverterFailure_IsNotRetried()
     {
         var newOptions = (TemporalClientOptions)Client.Options.Clone();
@@ -1769,7 +1706,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_CodecHandlerException_IsPassedThrough()
     {
         var newOptions = (TemporalClientOptions)Client.Options.Clone();
@@ -1838,7 +1774,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_ConverterHandlerException_IsPassedThrough()
     {
         var newOptions = (TemporalClientOptions)Client.Options.Clone();
@@ -1884,7 +1819,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_StartWorkflow_Succeeds()
     {
         // Build the worker options w/ the nexus service using the new generic handler
@@ -1907,7 +1841,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_Cancel_Succeeds()
     {
         // Build the worker options w/ the nexus service using the new generic handler
@@ -1941,7 +1874,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_SyncResult_Succeeds()
     {
         // Build the worker options w/ a handler that returns a sync result
@@ -1960,7 +1892,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_LinksAndContext_Populated()
     {
         // Capture the context and client passed to the generic handler so we can assert plumbing
@@ -2013,7 +1944,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_CancelOperation_CancelsUnderlying()
     {
         // The default CancelWorkflowRunAsync should cancel the underlying workflow when the
@@ -2051,7 +1981,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_CancelOverride_Invoked()
     {
         // Subclass with a CancelWorkflowRunAsync override; verify the override is used and
@@ -2091,7 +2020,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_StartWorkflowByName_Succeeds()
     {
         // Use the by-name overload of TemporalNexusClient.StartWorkflowAsync
@@ -2114,7 +2042,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_ConflictPolicy_UseExisting()
     {
         // Two operations with the same workflow ID + UseExisting policy attach to the same
@@ -2149,7 +2076,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_NoInputOverload_Succeeds()
     {
         // Exercise the no-input [TemporalOperation] path (2-arg method signature)
@@ -2196,7 +2122,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_StartActivity_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2232,7 +2157,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_StartActivityByName_Succeeds()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2259,7 +2183,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_CancelActivity_CancelsUnderlying()
     {
         ActivityStubs.WaitForCancelReached = new TaskCompletionSource();
@@ -2312,7 +2235,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_GenericHandler_CancelActivityOverride_Invoked()
     {
         ActivityStubs.WaitForCancelReached = new TaskCompletionSource();
@@ -2433,7 +2355,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_WorkflowStart()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2458,7 +2379,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_SyncResult()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2482,7 +2402,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_NoInput()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2522,7 +2441,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_VoidTypes()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2564,7 +2482,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_MixedWithHandlerFactory()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2580,70 +2497,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
                 "factory: hi",
                 await client.ExecuteNexusOperationAsync(svc => svc.ViaHandlerFactory("hi")));
         });
-    }
-
-    [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrBadReturnService
-    {
-        [TemporalOperation]
-        public Task<string> DoSomething(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, string input) =>
-            Task.FromResult(input);
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationBadReturnType_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrBadReturnService()));
-        Assert.Contains("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Contains(
-            "must return Task<TemporalOperationResult<String>>",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-    [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrBadParamsService
-    {
-        [TemporalOperation]
-        public Task<TemporalOperationResult<string>> DoSomething(string input) =>
-            Task.FromResult(TemporalOperationResult<string>.SyncResult(input));
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationBadParams_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrBadParamsService()));
-        Assert.Contains("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Contains(
-            "must accept parameters (TemporalOperationStartContext, ITemporalNexusClient, String)",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-#pragma warning disable CA1052 // Intentionally non-static so registration reaches signature check
-    [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrStaticService
-    {
-        [TemporalOperation]
-        public static Task<TemporalOperationResult<string>> DoSomething(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, string input) =>
-            Task.FromResult(TemporalOperationResult<string>.SyncResult(input));
-    }
-#pragma warning restore CA1052
-
-    [Fact]
-    public void AddNexusService_TemporalOperationStaticMethod_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrStaticService()));
-        Assert.Contains("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Contains(
-            "must not be static",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
     }
 
     [NexusService]
@@ -2663,7 +2516,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_CompositeGenericInput()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2675,68 +2527,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
                 ExecuteNexusOperationAsync(svc => svc.Sum(new List<int> { 1, 2, 3 }));
             Assert.Equal(6, result);
         });
-    }
-
-    [NexusServiceHandler(typeof(IGenericInputService))]
-    public class TemporalOperationAttrGenericInputMismatchService
-    {
-        [TemporalOperation]
-        public Task<TemporalOperationResult<int>> Sum(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, List<string> values) =>
-            Task.FromResult(TemporalOperationResult<int>.SyncResult(0));
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationGenericInputMismatch_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrGenericInputMismatchService()));
-        Assert.Contains("Failed obtaining operation handler from Sum", exc.Message);
-        Assert.Contains(
-            "must accept parameters",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-    [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrRawReturnService
-    {
-        [TemporalOperation]
-        public Task<int> DoSomething(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, string input) =>
-            Task.FromResult(0);
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationRawReturn_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrRawReturnService()));
-        Assert.Contains("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Contains(
-            "must return Task<TemporalOperationResult<String>>",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-    [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrDualAnnotationService
-    {
-        [TemporalOperation]
-        [NexusOperationHandler]
-        public Task<TemporalOperationResult<string>> DoSomething(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, string input) =>
-            Task.FromResult(TemporalOperationResult<string>.SyncResult(input));
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationDualAnnotation_Throws()
-    {
-        // The built-in [NexusOperationHandler] path claims the method first, and it
-        // rejects the signature (return type isn't IOperationHandler<,>).
-        Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrDualAnnotationService()));
     }
 
     [NexusServiceHandler(typeof(IStringService))]
@@ -2753,7 +2543,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_BindsToInstance()
     {
         var workerOptions = new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
@@ -2768,27 +2557,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [NexusServiceHandler(typeof(IStringService))]
-    public class TemporalOperationAttrPrivateService
-    {
-        [TemporalOperation]
-        private Task<TemporalOperationResult<string>> DoSomething(
-            TemporalOperationStartContext ctx, ITemporalNexusClient client, string input) =>
-            Task.FromResult(TemporalOperationResult<string>.SyncResult(input));
-    }
-
-    [Fact]
-    public void AddNexusService_TemporalOperationPrivateMethod_Throws()
-    {
-        var exc = Assert.Throws<ArgumentException>(() =>
-            new TemporalWorkerOptions($"tq-{Guid.NewGuid()}").
-                AddNexusService(new TemporalOperationAttrPrivateService()));
-        Assert.Contains("Failed obtaining operation handler from DoSomething", exc.Message);
-        Assert.Contains(
-            "must be public",
-            Assert.IsType<ArgumentException>(exc.InnerException).Message);
-    }
-
-    [NexusServiceHandler(typeof(IStringService))]
     public class TemporalOperationAttrThrowingService
     {
         [TemporalOperation]
@@ -2798,7 +2566,6 @@ public class NexusWorkerTests : WorkflowEnvironmentTestBase
     }
 
     [Fact]
-    [CloudTestExclusion(CloudTestExclusionReason.NeedsCloudAdaptation)]
     public async Task ExecuteNexusOperationAsync_TemporalOperationAttribute_ExceptionPropagatesUnwrapped()
     {
         // Compiled Expression call should propagate the user exception directly (no
